src/test/javascript/spec/app/entities/benefit/benefit.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { BenefitService } from 'app/entities/benefit/benefit.service';
import { IBenefit, Benefit } from 'app/shared/model/benefit.model';
import { BenefitType } from 'app/shared/model/enumerations/benefit-type.model';
import { RepetitionType } from 'app/shared/model/enumerations/repetition-type.model';

describe('Service Tests', () => {
  describe('Benefit Service', () => {
    let injector: TestBed;
    let service: BenefitService;
    let httpMock: HttpTestingController;
    let elemDefault: IBenefit;
    let expectedResult: IBenefit | IBenefit[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(BenefitService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Benefit(0, 'AAAAAAA', BenefitType.CASH, 0, 0, RepetitionType.NoRepetition);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Benefit', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Benefit()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Benefit', () => {
        const returnedFromService = Object.assign(
          {
            description: 'BBBBBB',
            type: 'BBBBBB',
            value: 1,
            frequency: 1,
            repetition: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Benefit', () => {
        const returnedFromService = Object.assign(
          {
            description: 'BBBBBB',
            type: 'BBBBBB',
            value: 1,
            frequency: 1,
            repetition: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Benefit', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
